requirements.txt
python-nmap
requests
scapy
geoip2
pdfkit
pytest
fastapi
uvicorn
httpx
reportlab
PyPDF2
apscheduler
pytest-benchmark
